Migrate to mongo 6

+++ Job/Data/MongoRepository.cs
@@ -1,18 +1,22 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com 
 
+using FtpClient;
 using Interfaces;
 using MongoDB.Bson;
 using MongoDB.Driver;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 
 namespace Job.Data
 {
-    public class MongoRepository : IRepository
+    public sealed class MongoRepository : IRepository
     {
         private IMongoDatabase _mongoDatabase;
+
+        public bool IsConnected { get; private set; }
 
         public void CreateConnection(string server, int port, string user, string password, string databaseName)
         {
@@ -24,17 +28,41 @@ namespace Job.Data
             };
 
             var client = new MongoClient(setting);
+           
+            
             _mongoDatabase = client.GetDatabase(databaseName);
+        }
+
+        private void Cluster_DescriptionChanged(object sender, MongoDB.Driver.Core.Clusters.ClusterDescriptionChangedEventArgs e)
+        {
+            IsConnected = e.NewClusterDescription.State == MongoDB.Driver.Core.Clusters.ClusterState.Connected;
+            Debug.WriteLine($"Mongodb description:{e.NewClusterDescription.State}");
+        }
+
+        public void CreateConnection(string connectingString,string databaseName)
+        {
+            var client = new MongoClient(connectingString);
+
+            Debug.WriteLine($"Mongodb description:{client.Cluster.Description.State}");
+            client.Cluster.DescriptionChanged += Cluster_DescriptionChanged;
+            
+            _mongoDatabase = client.GetDatabase(databaseName);
+
+            IsConnected = _mongoDatabase.RunCommandAsync((Command<BsonDocument>)"{ping:1}").Wait(1000);
         }
 
         public void Add<T>(T item) where T : class, new()
         {
+            if (!IsConnected) throw new Exception("Can`t add - Mongodb is offline"); 
+
             var col = _mongoDatabase.GetCollection<T>(typeof(T).Name);
             col.InsertOneAsync(item).Wait();
         }
 
         public void Delete<T>(T item) where T : IWithId
         {
+            if (!IsConnected) throw new Exception("Can`t delete - Mongodb is offline"); 
+
             var filter = Builders<T>.Filter.Eq(x => x.Id, item.Id);
             var col = _mongoDatabase.GetCollection<T>(typeof(T).Name);
             col.DeleteOneAsync(filter).Wait();
@@ -42,8 +70,11 @@ namespace Job.Data
 
         public List<T> All<T>() where T : class, new()
         {
+            if (!IsConnected) throw new Exception("Can`t get all - Mongodb is offline");
+
             var col = _mongoDatabase.GetCollection<T>(typeof(T).Name);
-            var j = col.FindAsync(o => true);
+            var filter = Builders<T>.Filter.Empty;
+            var j = col.FindAsync(filter);
             j.Wait();
             var res = j.Result.ToListAsync();
             res.Wait();
@@ -53,6 +84,8 @@ namespace Job.Data
 
         public void Update<T>(T item) where T : IWithId
         {
+            if (!IsConnected) throw new Exception("Can`t update - Mongodb is offline");
+
             var filter = Builders<T>.Filter.Eq(x => x.Id, item.Id);
             var col = _mongoDatabase.GetCollection<T>(typeof(T).Name);
             col.ReplaceOneAsync(filter, item).Wait();
@@ -60,25 +93,27 @@ namespace Job.Data
 
         public void Add<T>(string collection, T item) where T : class, new()
         {
+            if (!IsConnected) throw new Exception("Can`t add - Mongodb is offline");
+
             var col = _mongoDatabase.GetCollection<T>(collection);
             col.InsertOneAsync(item).Wait();
         }
 
         public void Delete<T>(string collection, T item) where T : IWithId
         {
+            if (!IsConnected) throw new Exception("Can`t delete - Mongodb is offline"); 
+
             var filter = Builders<T>.Filter.Eq(x => x.Id, item.Id);
             var col = _mongoDatabase.GetCollection<T>(collection);
             col.DeleteOneAsync(filter).Wait();
         }
 
-        public List<T> All<T>(string collection) where T : class, new()
+        public List<T> All<T>(string collection) where T : class,  new()
         {
-            //.Where(x => _profile.StatusManager.IsViewStatusChecked(x.StatusCode));
+            if (!IsConnected) throw new Exception("Can`t get all - Mongodb is offline"); 
 
             var col = _mongoDatabase.GetCollection<T>(collection);
-            //var j =  col.FindAsync(o => true);
-            var j = col.FindAsync(o => true);
-
+            var j = col.FindAsync(new BsonDocument());
             j.Wait();
             var res = j.Result.ToListAsync();
             res.Wait();
@@ -87,11 +122,15 @@ namespace Job.Data
 
         public object GetRawCollection<T>(string collection) where T : class, new()
         {
+            if (!IsConnected) throw new Exception("Can`t GetRawCollection - Mongodb is offline");
+
             return _mongoDatabase.GetCollection<T>(collection);
         }
 
         public void Update<T>(string collection, T item) where T : IWithId
         {
+            if (!IsConnected) throw new Exception("Can`t update - Mongodb is offline");
+
             var filter = Builders<T>.Filter.Eq(x => x.Id, item.Id);
             var col = _mongoDatabase.GetCollection<T>(collection);
             col.ReplaceOneAsync(filter, item).Wait();
@@ -99,6 +138,8 @@ namespace Job.Data
 
         public T GetById<T>(string collection, object id) where T : class, IWithId
         {
+
+            if (!IsConnected) throw new Exception("Can`t get by id - Mongodb is offline");
 
             ObjectId objectId = ObjectId.Empty;
 
@@ -116,7 +157,10 @@ namespace Job.Data
             }
 
             var col = _mongoDatabase.GetCollection<T>(collection);
-            var f = col.FindAsync(x => x.Id.Equals(objectId));
+
+            var filter = Builders<T>.Filter.Eq(t=>t.Id, objectId);
+
+            var f = col.FindAsync(filter);
             f.Wait();
             var l = f.Result.ToListAsync(default);
             l.Wait();
@@ -126,6 +170,13 @@ namespace Job.Data
             }
 
             return null;
+        }
+
+        public object GetRawCollection<T>() where T : class, new()
+        {
+            if (!IsConnected) throw new Exception("Can`t GetRawCollection - Mongodb is offline");
+            var name = typeof(T).Name;
+            return _mongoDatabase.GetCollection<T>(name);
         }
     }
 }
